create read only signal
